UI Completed
Refactored and started the idea of migrating UI Components out to a separate library. This would help with scalability and consistency later on.

Refactoring caused us to lose out on the _onBlur_ and _onSubmit_ formatting of the number, it still gets formatted for the call, but not in the UI.